added timer, deleted custom hook (handler logout after click), handler logout moved to timer component, addition, when timer reaches value equale 0, logout will be perfomed
