+++ src/App.tsx
@@ -15,7 +15,6 @@ import { BrowserRouter,
 import Cookies from 'universal-cookie';
 
 import { getAccount } from 'helpers/account.helper';
-import { useHandlerLogout } from 'helpers/hooks/useHandlerLogout';
 import { Decrypt, DecryptCookie } from 'helpers/password.helper';
 import { reduxIsAuthentication, reduxSetRoles, reduxSetTeam } from 'store/actions/user';
 
@@ -42,7 +41,8 @@ const App = (): JSX.Element => {
   // const [roles, setRoles] = useState<string[] | null>(null);
   // const [team, setTeam] = useState<string | null>(null);
   const [redirectToLogin, setRedirectToLogin] = useState<boolean>(false);
-  
+
+
   useEffect(() => {
     const downloadData = async () => {
       await getAccountState();
@@ -68,7 +68,6 @@ const App = (): JSX.Element => {
     };
     dataLogin ? checkLogin(Decrypt(dataLogin.login), dataLogin.password) : setRedirectToLogin(true);
   },[]);
-  useHandlerLogout();
 
   const DashBoard = React.lazy(() => import( './pages/DashBoard/Dashboard'));
   const Codes = React.lazy(() => import( './pages/Codes/Codes'));
