+++ src/constans/Constans.ts
@@ -1,0 +1,1 @@
+export const timeToLogout = 1800;
